Move tile creation from MapsforgeHandler to MapsforgeTaskHandler
