+++ src/main/java/com/telemaxx/mapsforgesrv/MapsforgeTaskHandler.java
@@ -1,6 +1,10 @@
 package com.telemaxx.mapsforgesrv;
 
+import java.awt.image.BufferedImage;
+import java.awt.image.DataBufferInt;
+import java.io.ByteArrayOutputStream;
 import java.io.File;
+import java.io.IOException;
 import java.io.InputStream;
 import java.nio.channels.FileChannel;
 import java.nio.file.FileSystem;
@@ -13,7 +17,15 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Set;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
+import javax.imageio.ImageIO;
+
+import org.eclipse.jetty.server.Request;
+import org.mapsforge.core.graphics.TileBitmap;
+import org.mapsforge.core.model.Tile;
+import org.mapsforge.map.awt.graphics.AwtGraphicFactory;
 import org.mapsforge.map.datastore.MultiMapDataStore;
 import org.mapsforge.map.layer.cache.TileCache;
 import org.mapsforge.map.layer.hills.DemFolderFS;
@@ -25,6 +37,7 @@ import org.mapsforge.map.layer.hills.SimpleShadingAlgorithm;
 import org.mapsforge.map.layer.labels.TileBasedLabelStore;
 import org.mapsforge.map.layer.renderer.DatabaseRenderer;
 import org.mapsforge.map.layer.renderer.DirectRenderer;
+import org.mapsforge.map.layer.renderer.RendererJob;
 import org.mapsforge.map.model.DisplayModel;
 import org.mapsforge.map.reader.MapFile;
 import org.mapsforge.map.rendertheme.ExternalRenderTheme;
@@ -38,6 +51,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.github.marschall.memoryfilesystem.MemoryFileSystemBuilder;
+
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletOutputStream;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 public class MapsforgeTaskHandler {
 	
@@ -64,6 +82,10 @@ public class MapsforgeTaskHandler {
 	private Map<String, DirectRenderer> directRenderer = null;
 
 	private MapsforgeHandler mapsforgeHandler;
+	private MapsforgeConfig mapsforgeConfig;
+	private MapsforgeTaskConfig mapsforgeTaskConfig;
+	
+	private static final Pattern requestPathPattern = Pattern.compile("/(\\d+)/(-?\\d+)/(-?\\d+)(?:(?:\\.)(.*))?"); //$NON-NLS-1$
 	
 	public MapsforgeTaskHandler(MapsforgeHandler mapsforgeHandler, MapsforgeTaskConfig mapsforgeTaskConfig, String name) throws Exception {
 		
@@ -71,8 +93,9 @@ public class MapsforgeTaskHandler {
 		
 		this.name = name;
 		this.mapsforgeHandler = mapsforgeHandler;
-		MapsforgeConfig mapsforgeConfig = mapsforgeHandler.getMapsforgeConfig();
-		
+		this.mapsforgeConfig = mapsforgeHandler.getMapsforgeConfig();
+		this.mapsforgeTaskConfig = mapsforgeTaskConfig;
+
 		logger.info("################### MAPS INFO ###################");
 		multiMapDataStore = new MultiMapDataStore(MultiMapDataStore.DataPolicy.RETURN_ALL);
 		if (mapsforgeTaskConfig.getMapFiles().size() == 0) {
@@ -273,6 +296,192 @@ public class MapsforgeTaskHandler {
 	    }
 		new Thread(null,renderThemeFuture,tname).start();
 	}
+	
+	protected void handle(String target, Request baseRequest, HttpServletRequest request, HttpServletResponse response) throws Exception {
+		String path = request.getPathInfo();
+		String engine = "std";
+		
+		int x, y, z;
+		String ext = MapsforgeConfig.TILE_EXTENSION; // $NON-NLS-1$
+		Matcher m = requestPathPattern.matcher(path);
+		if (m.matches()) {
+			x = Integer.parseInt(m.group(2));
+			y = Integer.parseInt(m.group(3));
+			z = Integer.parseInt(m.group(1));
+			if (m.group(4) != null) ext = m.group(4);
+		} else {
+			try {
+				x = Integer.parseInt(request.getParameter("x")); //$NON-NLS-1$
+				y = Integer.parseInt(request.getParameter("y")); //$NON-NLS-1$
+				z = Integer.parseInt(request.getParameter("z")); //$NON-NLS-1$
+			} catch (NumberFormatException e) {
+				logger.error("Invalid tile request "+path); //$NON-NLS-1$
+				response.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
+				return;
+			};
+		}
+		if (x < 0 || x >= (1 << z)) {
+			logger.error("Tile number x=" + x + " out of range!"); //$NON-NLS-1$
+			response.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
+			return;
+		}
+		if (y < 0 || y >= (1 << z)) {
+			logger.error("Tile number y=" + y + " out of range!"); //$NON-NLS-1$
+			response.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
+			return;
+		}
+
+		float requestedTextScale = 1.0f; // Original text scaling comes from config value
+		try {
+			String tmp = request.getParameter("textScale"); //$NON-NLS-1$
+			if (tmp != null) {
+//				Override text scaling from config value by text scaling from HTTP request
+//				Final text scaling = textScale * requestedTextScale
+				requestedTextScale = Float.parseFloat(tmp) / DisplayModel.textScale;
+			}
+		} catch (Exception e) {
+			throw new ServletException("Failed to parse \"textScale\" property: " + e.getMessage(), e); //$NON-NLS-1$
+		}
+
+		float requestedUserScale = mapsforgeTaskConfig.getUserScale();
+		try {
+			String tmp = request.getParameter("userScale"); //$NON-NLS-1$
+			if (tmp != null) {
+				requestedUserScale = Float.parseFloat(tmp);
+			}
+		} catch (Exception e) {
+			throw new ServletException("Failed to parse \"userScale\" property: " + e.getMessage(), e); //$NON-NLS-1$
+		}
+
+		boolean requestedTransparent = MapsforgeConfig.DEFAULT_TRANSPARENT;
+		try {
+			String tmp = request.getParameter("transparent"); //$NON-NLS-1$
+			if (tmp != null) {
+				requestedTransparent = Boolean.parseBoolean(tmp);
+			}
+		} catch (Exception e) {
+			throw new ServletException("Failed to parse \"transparent\" property: " + e.getMessage(), e); //$NON-NLS-1$
+		}
+
+		int requestedTileRenderSize = MapsforgeConfig.DEFAULT_TILE_RENDERSIZE;
+		try {
+			String tmp = request.getParameter("tileRenderSize"); //$NON-NLS-1$
+			if (tmp != null)
+				requestedTileRenderSize = Integer.parseInt(tmp);
+		} catch (Exception e) {
+			throw new ServletException("Failed to parse \"tileRenderSize\" property: " + e.getMessage(), e); //$NON-NLS-1$
+		}
+
+		TileBitmap tileBitmap = null;
+		Tile tile = new Tile(x, y, (byte) z, requestedTileRenderSize);
+		if (multiMapDataStore.supportsTile(tile)) {
+			boolean enable_hs = true;
+			try {
+				String tmp = request.getParameter("hillshading"); //$NON-NLS-1$
+				if (tmp != null)
+					enable_hs = Integer.parseInt(request.getParameter("hillshading")) != 0; //$NON-NLS-1$
+			} catch (Exception e) {
+				throw new ServletException("Failed to parse \"hillshading\" property: " + e.getMessage(), e); //$NON-NLS-1$
+			}
+			if (hillsRenderConfig != null && enable_hs) engine = "hs";
+
+			// requestedUserScale = 2.0f; // Uncomment for testing purpose only!
+//			Calling "displayModel.setUserScaleFactor" alone has no visible impact on rendering.
+//			Starting new "renderThemeFuture" afterwards helps, but can significantly impact rendering performance
+//			if different TMS clients take turns requesting different userScale values !!!
+			if (displayModel.getUserScaleFactor() != requestedUserScale) {
+				displayModel.setUserScaleFactor (requestedUserScale);
+				updateRenderThemeFuture();
+			}
+
+			RendererJob job = new RendererJob(tile, multiMapDataStore, renderThemeFuture, displayModel,
+				requestedTextScale, requestedTransparent, false);
+
+//Synchronizing render jobs has no visible effect -> disabled
+//				synchronized (this) {
+				if (directRenderer != null) {
+					tileBitmap = directRenderer.get(engine).executeJob(job);
+				} else {
+					tileBitmap = databaseRenderer.get(engine).executeJob(job);
+				}
+				if (!hillShadingOverlay) tileCache.put(job, null);
+//				}
+		}
+
+		BufferedImage image;
+		if (tileBitmap != null) {
+			image = AwtGraphicFactory.getBitmap(tileBitmap); // image type is TYPE_INT_RGB
+			//int imageType = image.getType(); // returns 1 (TYPE_INT_RGB)
+			int imageWidth  = image.getWidth();
+			int imageHeight = image.getHeight();
+			//int dataBufferType = image.getRaster().getDataBuffer().getDataType(); // returns 3 (TYPE_INT)
+			// DataBuffer created by Mapsforge renderer is of type DataBufferInt,
+			// i.e. one int value 0xaarrggbb per pixel
+			DataBufferInt dataBuffer = (DataBufferInt) image.getRaster().getDataBuffer();
+			int[] pixelArray = dataBuffer.getData();
+			int pixelValue;
+			int pixelCount = imageWidth * imageHeight;
+			if (hillShadingOverlay) { // transparent hillshading overlay image requested
+				BufferedImage newImage = new BufferedImage (imageWidth, imageHeight, BufferedImage.TYPE_INT_ARGB);
+				DataBufferInt newDataBuffer = (DataBufferInt) newImage.getRaster().getDataBuffer();
+				int[] newPixelArray = newDataBuffer.getData();
+				while (pixelCount-- > 0) {
+					pixelValue = pixelArray[pixelCount];
+					if (pixelValue == 0xfff8f8f8) { // 'nodata' hillshading value
+						newPixelArray[pixelCount] = 0x00000000; // fully transparent pixel
+					} else { // get gray value of pixel from blue value of pixel
+						newPixelArray[pixelCount] = colorLookupTable[pixelValue & 0xff];
+					}
+				}
+				image = newImage; // return transparent overlay image instead of original image
+			} else if (colorLookupTable != null) { // apply gamma correction and/or contrast-stretching
+				while (pixelCount-- > 0) {
+					pixelValue = pixelArray[pixelCount];
+					pixelArray[pixelCount] = (pixelValue & 0xff000000) // alpha value
+							| (colorLookupTable[(pixelValue >>> 16) & 0xff] << 16) // red value
+							| (colorLookupTable[(pixelValue >>> 8) & 0xff] << 8) // green value
+							| (colorLookupTable[pixelValue & 0xff]); // blue value
+				}
+			}
+			response.setStatus(HttpServletResponse.SC_OK);
+		} else {
+			String outOfRangeTms = mapsforgeConfig.getOutOfRangeTms();;
+			if(outOfRangeTms != null) {
+				response.setStatus(HttpServletResponse.SC_MOVED_TEMPORARILY);
+				String redirecturl = outOfRangeTms.replace("{x}", x+"").replace("{y}", y+"").replace("{z}", z+"");
+				response.setHeader("Location", redirecturl);
+				response.flushBuffer();
+				logger.info("out-of-range redirect '"+redirecturl+"'");
+				return;
+			} else {
+				image = MapsforgeConfig.BI_NOCONTENT;
+				response.setStatus(HttpServletResponse.SC_NOT_FOUND);
+			}
+		}
+		if (mapsforgeConfig.getCacheControl() > 0) {
+			response.addHeader("Cache-Control", "public, max-age=" + mapsforgeConfig.getCacheControl()); //$NON-NLS-1$ //$NON-NLS-2$
+		}
+		response.setContentType("image/" + ext); //$NON-NLS-1$
+		//ImageIO.write(image, ext, response.getOutputStream());
+		int bufferSize = 256 + 4*image.getWidth()*image.getHeight(); // Assume image data size <= bufferSize
+		MyResponseBufferOutputStream responseBufferStream = new MyResponseBufferOutputStream(bufferSize);
+		ImageIO.write(image, ext, responseBufferStream);
+		responseBufferStream.flush(response);
+		responseBufferStream.close();
+	}
+
+	// Extend class ByteArrayOutputStream by setting content length on buffer flush
+	private static class MyResponseBufferOutputStream extends ByteArrayOutputStream {
+		public MyResponseBufferOutputStream(int bufferSize) {
+			buf = new byte[bufferSize];
+		}
+		public void flush (HttpServletResponse response) throws IOException {
+			response.setContentLength(count);
+			ServletOutputStream responseOutputStream = response.getOutputStream();
+			responseOutputStream.write(buf, 0, count);
+			responseOutputStream.flush();
+		}
+	}
 
 	/**
 	 * displaying the containing styles inside theme xml if a style is given(!=null)
@@ -300,46 +509,6 @@ public class MapsforgeTaskHandler {
 		if (selectedStyleExists && themeFileStyle != null) {
 			logger.info("Defined style '"+themeFileStyle+"' used"); //$NON-NLS-1$
 		}
-	}
-
-	public MultiMapDataStore getMultiMapDataStore() {
-		return multiMapDataStore;
-	}
-	
-	public boolean getHillShadingOverlay() {
-		return hillShadingOverlay;
-	}
-	
-	public HillsRenderConfig getHillsRenderConfig() {
-		return hillsRenderConfig;
-	}
-	
-	public DisplayModel getDisplayModel() {
-		return displayModel;
-	}
-
-	public RenderThemeFuture getRenderThemeFuture() {
-		return renderThemeFuture;
-	}
-	
-	public TileCache getTileCache() {
-		return tileCache;
-	}
-	
-	public String getName() {
-		return name;
-	}
-
-	public int[] getColorLookupTable() {
-		return colorLookupTable;
-	}
-	
-	public Map<String, DatabaseRenderer> getDatabaseRenderer() {
-		return databaseRenderer;
-	}
-
-	public Map<String, DirectRenderer> getDirectRenderer() {
-		return directRenderer;
 	}
 	
 	// Enumeration of tile server's internal rendering themes
